Add IP-based authorization to EPCR issues and IP helper route
A new function 'AuthEpcrIssueModule' is added to the 'auth' handler, creating an IP-based authorization for EPCR issue routes. It reads an authorization IP list from the server environment and checks if the client IP is authorized. Also, a new helper route 'ip' is added to return the client's IP address.