Add news feed functionality
This update adds a new feature to manage and display a news feed. It includes the creation of a database table, endpoints, as well as the ability to retrieve and display all news. We also have provisions for a watcher to monitor and update changes in the news feed, ensuring that it is updated in real time.