Add news feed updates to websocket broadcast
A new endpoint "/newsfeed" was added to the websocket router to allow updates on the news feed. The "NewsUpdate" function was defined in the websocket handler to handle news feed messages. A "News feed Event" constant was also created to clearly mark these events. Now, news feed updates are also broadcasted to all the connected clients.